Completing username and password validation and also adding empty fie…

+++ client/src/components/login.js
@@ -10,20 +10,25 @@ export default class Login extends Component {
     }
 
     handleSignIn = () => {
-        console.log("Sending user info for sign in");
-        login(this.state)
-        .then(res => {
-            if (res.data === null) {
-                localStorage.clear();
-                alert("Username or Password is incorrect!");
-                this.setState({email: "", password: ""});
-            }
-            else {
-                localStorage.setItem("currentUser", JSON.stringify(res.data))
-                this.props.setUser(res.data);
-                window.location.replace("/member")
-            }
-        });
+        if (this.state.email && this.state.password) {
+            login(this.state)
+            .then(res => {
+                if (res.data === null) {
+                    localStorage.clear();
+                    alert("Username or Password is incorrect!");
+                    this.setState({email: "", password: ""});
+                }
+                else {
+                    localStorage.setItem("currentUser", JSON.stringify(res.data))
+                    this.props.setUser(res.data);
+                    window.location.replace("/member")
+                }
+            });
+        }
+        else {
+            alert("Please fill up all the fields");
+            this.setState({email: "", password: ""});
+        }
     }
 
     handleChange = ({target:{name,value}}) => this.setState({[name]:value})
